pyproject.toml
[tool.poetry]
name = "text-search"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.1"
pydantic-settings = "^2.4.0"
elasticsearch = "^8.15.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
